Postgres ITs to fix

+++ application/rest-api-adapter/src/main/java/catlean/monolithic/backend/rest/api/adapter/TeamRestApiAdapter.java
@@ -32,7 +32,7 @@ public class TeamRestApiAdapter implements TeamApi {
         try {
             final User authenticatedUser = authenticationService.getAuthenticatedUser();
 
-            final Map<String, List<Integer>> repositoryIdsMappedToTeamName =
+            final Map<String, List<Long>> repositoryIdsMappedToTeamName =
                     getRepositoryIdsMappedToTeamName(createTeamRequestContract);
             final List<Team> teamsForNameAndRepositoriesAndUser =
                     teamFacadeAdapter.createTeamsForNameAndRepositoriesAndUser(repositoryIdsMappedToTeamName,
