+++ domain/src/main/java/fr/catlean/monolithic/backend/domain/service/account/TeamService.java
@@ -18,7 +18,7 @@ public class TeamService implements TeamFacadeAdapter {
     private final AccountTeamStorage accountTeamStorage;
 
     @Override
-    public List<Team> createTeamsForNameAndRepositoriesAndUser(final Map<String, List<Integer>> repositoryIdsMappedToTeamName,
+    public List<Team> createTeamsForNameAndRepositoriesAndUser(final Map<String, List<Long>> repositoryIdsMappedToTeamName,
                                                                User user) throws CatleanException {
         user.hasConfiguredTeam();
         final List<Team> teams = new ArrayList<>();
